backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Web scraping
scrapy==2.11.0
selenium==4.15.2
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
aiohttp==3.9.1

# LLM and AI
# openai==1.3.7  # Removed (using Perplexity instead)
# langchain==0.0.350  # Removed
# sentence-transformers==2.2.2  # Removed
pgvector==0.2.4

# Background tasks and caching
celery==5.3.4
redis==5.0.1

# Email notifications
sendgrid==6.11.0

# Utilities
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-dotenv==1.0.0
loguru==0.7.2
tenacity==8.2.3

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Date and time handling
python-dateutil==2.8.2
pytz==2023.3

# Web driver management
webdriver-manager==4.0.1 